First working test
Nor shapes or waves are editable yet. As of now, the shapes are decomposed in lines and arches, then those are segmented depending on the wave used. A piece of the wave is drawn on every segment. Something is not right in the segments that make the transition between lines and arches, not sure what it is.